+++ ci4/app/Entities/AutoUpdate.php
@@ -1,7 +1,6 @@
 <?php namespace App\Entities;
 
-use App\Libraries\Kubernetes\KubeHelper;
-use App\Libraries\PodioLib;
+use App\Libraries\PostUpdateActions\PostUpdateActionHelper;
 use App\Libraries\ZMQ\ChangeEvent;
 use App\Libraries\ZMQ\Events;
 use App\Libraries\ZMQ\ZMQProxy;
@@ -66,19 +65,13 @@ class AutoUpdate extends Entity {
         $deployment->find($this->deployment_id);
         $deployment->updateVersion($this->next_tag);
 
-        if ($deployment->enable_podio_notification && $deployment->workspace_id) {
-            $spec = $this->deployment->findDeploymentSpecification();
-            if (!$spec->container_image->exists()) {
-                $spec->container_image->find();
-            }
+        $postUpdateActionHelper = new PostUpdateActionHelper($deployment);
+        $postUpdateActionHelper->performAll();
 
-            try {
-                $podioLib = new PodioLib($spec->container_image);
-                $podioLib->notify($deployment, fn($log) => $this->appendLog($log));
-            } catch (\Exception $e) {
-                $this->appendLog('Failed to notify podio');
-                $this->appendLog(KubeHelper::PrintException($e));
-            }
+        $log = Data::getDebugger();
+        if (is_array($log)) {
+            $lines = implode("\n", $log);
+            $this->appendLog($lines);
         }
 
         ZMQProxy::getInstance()->send(