+++ ci4/app/Entities/PodioIntegration.php
@@ -1,5 +1,7 @@
 <?php namespace App\Entities;
 
+use App\Interfaces\PodioIntegrationGetFieldsResponse;
+use DebugTool\Data;
 use RestExtension\Core\Entity;
 
 /**
@@ -15,6 +17,48 @@ use RestExtension\Core\Entity;
  * @property PodioFieldReference $podio_field_references
  */
 class PodioIntegration extends Entity {
+
+    /**
+     * @return PodioIntegrationGetFieldsResponse[]
+     */
+    public function getFields(): array {
+        $client = new \PodioClient($this->client_id, $this->client_secret);
+        $client->authenticate_with_app($this->app_id, $this->app_token);
+        $app = \PodioApp::get($client, $this->app_id);
+        $fields = [];
+        foreach ($app->fields as $field) {
+            Data::debug($field->config);
+            $fields[] = [
+                'id' => (string)$field->id,
+                'name' => $field->label,
+                'type' => $field->type,
+            ];
+        }
+        return $fields;
+    }
+
+    public function getFieldDetails(string $fieldId): array {
+        $client = new \PodioClient($this->client_id, $this->client_secret);
+        $client->authenticate_with_app($this->app_id, $this->app_token);
+        $appField = \PodioAppField::get($client, $this->app_id, $fieldId);
+
+        return [
+            'id' => (string)$appField->id,
+            'name' => $appField->name,
+            'type' => $appField->type,
+            'options' => array_map(
+                fn($option) => [
+                    'id' => (string)$option['id'],
+                    'text' => (string)$option['text'],
+                    'color' => (string)$option['color'],
+                ],
+                array_values(array_filter(
+                    $appField->config['settings']['options'] ?? [],
+                    fn($option) => $option['status'] == 'active'
+                ))
+            )
+        ];
+    }
 
     /**
      * @return \ArrayIterator|\OrmExtension\Extensions\Entity[]|\Traversable|PodioIntegration[]