+++ ci4/app/Interfaces/PodioIntegrationGetFieldDetailsResponse.php
@@ -1,0 +1,13 @@
+<?php namespace App\Interfaces;
+
+/**
+ * Interface PodioIntegrationGetFieldDetailsResponse
+ * @package App\Interfaces
+ * @property string $name
+ * @property string $id
+ * @property string $type
+ * @property PodioIntegrationGetFieldDetailsResponseOption[] $options
+ */
+interface PodioIntegrationGetFieldDetailsResponse {
+
+}