+++ ci4/app/Interfaces/PodioIntegrationGetFieldDetailsResponseOption.php
@@ -1,0 +1,12 @@
+<?php namespace App\Interfaces;
+
+/**
+ * Interface PodioIntegrationGetFieldDetailsResponseOption
+ * @package App\Interfaces
+ * @property string $color
+ * @property string $id
+ * @property string $text
+ */
+interface PodioIntegrationGetFieldDetailsResponseOption {
+
+}