+++ ci4/app/Interfaces/PostUpdateActionConditionList.php
@@ -1,7 +1,9 @@
 <?php namespace App\Interfaces;
+use App\Entities\PostUpdateActionCondition;
+
 /**
  * @property PostUpdateActionCondition[] $values
  */
-interface PostUpdateActionList {
+interface PostUpdateActionConditionList {
 
 }